requirements.txt
annotated-types==0.6.0
anyio==4.3.0
Authlib==1.3.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.4
filelock==3.13.1
Flask==3.0.2
fsspec==2024.2.0
google-api-core==2.17.1
google-auth==2.28.1
google-cloud-speech==2.25.0
google-cloud-texttospeech==2.16.2
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-health-checking==1.62.0
grpcio-status==1.62.0
grpcio-tools==1.62.0
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
packaging==23.2
pillow==10.2.0
profanity-check==1.0.3
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.3.1
sentencepiece==0.2.0
sniffio==1.3.0
sympy==1.12
textblob==0.18.0.post0
threadpoolctl==3.3.0
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.38.1
typing_extensions==4.9.0
urllib3==2.2.1
validators==0.22.0
weaviate-client==4.4.4
Werkzeug==3.0.1
